pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "deepresearch"
version = "0.1.0"
description = "Automated AI Research Pipeline"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "DeepResearch Team" }
]
keywords = ["ai", "research", "automation", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # AI Providers
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "google-generativeai>=0.8.0",
    # HTTP & Async
    "httpx>=0.27.0",
    "aiofiles>=24.0.0",
    # Literature Search
    "arxiv>=2.1.0",
    # Data & Analysis
    "numpy>=1.26.0",
    "scipy>=1.12.0",
    "pandas>=2.2.0",
    # Visualization
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    # Configuration
    "hydra-core>=1.3.0",
    "pydantic>=2.0.0",
    # CLI
    "typer>=0.12.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

[project.scripts]
deepresearch = "deepresearch.cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src/deepresearch"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
